Cloud Computing/Model API/requirements.txt
Flask
scikit-learn
tensorflow
pandas
numpy
PyJWT
python-dotenv
